+++ src/SyncServiceProvider.php
@@ -9,6 +9,7 @@ use Override;
 use Vitamin2\Sync\Commands\SyncBackupsCleanCommand;
 use Vitamin2\Sync\Commands\SyncCommand;
 use Vitamin2\Sync\Commands\SyncCommandsCommand;
+use Vitamin2\Sync\Commands\SyncDoctorCommand;
 use Vitamin2\Sync\Commands\SyncListCommand;
 use Vitamin2\Sync\Commands\SyncTestConnectionCommand;
 
@@ -44,6 +45,7 @@ class SyncServiceProvider extends ServiceProvider
             SyncCommandsCommand::class,
             SyncBackupsCleanCommand::class,
             SyncTestConnectionCommand::class,
+            SyncDoctorCommand::class,
         ]);
     }
 }